wss_frame_write: Fix extended length for frames larger than 65,535 bytes.
The call to htonl() in wss_frame_write does not fill the entire 8 bytes
allocate in the frame for it, which led to use of uninitialized memory
in frames using the extended payload (small length of 127).

This is fixed by setting the highest order byte to 0 as is directed
in the specification.

Fixes: #1