wss.c: Fix parsing of 64-bit numbers.
If the initial nominal length is 127 (large extended
length frame), we would read the wrong length because
ntohl operates on 32 bits, not 64 bits. For example,
if the higher order bits were all 0, then it would
return 0.

Using ntohl is incorrect because that is a
32-bit function. We now use a 64-bit network
to host function to convert the extended
length properly.